update JavaBasics to read radius as an integer instead of a float

+++ DSA_Java/JavaBasics.java
@@ -2,7 +2,7 @@ import java.util.*;
 public class JavaBasics{
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        float rad = sc.nextFloat();
+        float rad = sc.nextInt();
         float area = 3.14f * rad * rad;
         System.out.println("Area: "+area);
     }
